deletion of files

+++ backend/controllers/advertiserController.js
@@ -40,6 +40,22 @@ const deleteAdvertiserAccount = async (req, res) => {
       await cloudinary.uploader.destroy(advertiser.logo.public_id);
     }
 
+    const gfs = req.app.locals.gfs;
+    if (!gfs) {
+      return res.status(500).send("GridFS is not initialized");
+    }
+    const fileNames = [
+      advertiser.files.IDFilename,
+      advertiser.files.taxationRegistryCardFilename,
+    ];
+    const files = await gfs.find({ filename: { $in: fileNames } }).toArray();
+    if (!files || files.length === 0) {
+      return res.status(404).json({ err: "No file exists" });
+    }
+    files.forEach(async (file) => {
+      await gfs.delete(file._id);
+    });
+
     await Advertiser.findByIdAndDelete(res.locals.user_id);
 
     res
